+++ modules/dialog/dialog-factory.js
@@ -9,6 +9,5 @@ export const HMDialogFactory = (dataset, caller = null, opt = {}) => {
     if (dialog === "def") return HMDialog.getDefendDialog(dataset, caller);
     if (dialog === "dmg") return HMDialog.getDamageDialog(dataset, caller);
     if (dialog === "save") return HMDialog.getSaveDialog(dataset, caller);
-    if (dialog === "skill") return HMDialog.getSkillDialog(dataset, caller);
     return undefined;
 };