+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.5.7] - 2025-08-15
+
+### Added
+
+- Bulk skill roll chat card.
+
+### Changed
+
+- Refreshed skill dialog.
+
+### Removed
+
+- Legacy skill macros.
+
 ## [0.5.6] - 2025-08-12
 
 ### Fixed